Confirmação de remoção

+++ app/src/main/java/com/williamnery/convidados/view/viewholder/GuestViewHolder.kt
@@ -1,6 +1,8 @@
 package com.williamnery.convidados.view.viewholder
 
+import android.content.DialogInterface
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.williamnery.convidados.databinding.RowGuestBinding
 import com.williamnery.convidados.model.GuestModel
@@ -17,7 +19,15 @@ class GuestViewHolder(private val bind: RowGuestBinding, private val listener: O
         }
 
         bind.textName.setOnLongClickListener {
-            listener.onDelete(guest.id)
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Remoção de convidado")
+                .setMessage("Tem certeza que deseja remover?")
+                .setPositiveButton("Sim") { dialog, which ->
+                    listener.onDelete(guest.id)
+                }
+                .setNegativeButton("Não", null)
+                .create()
+                .show()
             true
         }
     }
